all: Refactor Packet Broker gateway identifiers
