+++ pkg/packetbroker/ids.go
@@ -1,4 +1,4 @@
-// Copyright © 2019 The Things Network Foundation, The Things Industries B.V.
+// Copyright © 2023 The Things Network Foundation, The Things Industries B.V.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,9 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package cluster
+package packetbroker
 
 import "go.thethings.network/lorawan-stack/v3/pkg/ttnpb"
 
-// PacketBrokerGatewayID is the proxy gateway identifier of gateways connected through Packet Broker.
-var PacketBrokerGatewayID = &ttnpb.GatewayIdentifiers{GatewayId: "packetbroker"}
+// GatewayIdentifiers are the proxy gateway identifier of gateways connected through Packet Broker.
+var GatewayIdentifiers = &ttnpb.GatewayIdentifiers{GatewayId: "packetbroker"}
